+++ src/ts/Ads/Utilities/ProgrammaticTrackingService.ts
@@ -20,7 +20,10 @@ export enum ProgrammaticTrackingError {
 export enum AdmobMetric {
     AdmobUsedCachedVideo = 'admob_used_cached_video',
     AdmobUsedStreamedVideo = 'admob_used_streamed_video',
-    AdmobUserVideoSeeked = 'admob_user_video_seeked'
+    AdmobUserVideoSeeked = 'admob_user_video_seeked',
+    AdmobRewardedVideoStart = 'admob_rewarded_video_start',
+    AdmobUserWasRewarded = 'admob_user_was_rewarded',
+    AdmobUserSkippedRewardedVideo = 'admob_user_skipped_rewarded_video'
 }
 
 export enum BannerMetric {